+++ packages/sdk/e2e/tests/mobile/executors/tts-executor.ts
@@ -48,14 +48,14 @@ export class MobileTtsExecutor extends ModelAssetExecutor<typeof ttsTests> {
     return makeEnhancedTtsHandler<Expectation, TestResult>({
       dependency: dep,
       ensureLoaded: (dependency) => this.resources.ensureLoaded(dependency),
-      validate: ValidationHelpers.validate
+      validate: (output, expectation) => ValidationHelpers.validate(output, expectation)
     })
   }
 
   private makeOutputSampleRateComparison() {
     return makeOutputSampleRateComparisonHandler<Expectation, TestResult>({
       ensureLoaded: (dependency) => this.resources.ensureLoaded(dependency),
-      validate: ValidationHelpers.validate
+      validate: (output, expectation) => ValidationHelpers.validate(output, expectation)
     })
   }